Compress all-zero time-series metrics
