+++ src/visualizer.rs
@@ -1,5 +1,6 @@
 use crate::data::data_formats::{AperfData, DataFormat, ProcessedData};
-use crate::utils::{combine_value_ranges, topological_sort};
+use crate::data::utils::compress_all_zero_time_series_metric;
+use crate::data::utils::{combine_value_ranges, topological_sort};
 use crate::{data::Data, data::ReportData, get_file};
 use anyhow::Result;
 use log::{debug, error};
@@ -149,6 +150,8 @@ impl DataVisualizer {
 ///   graphs could have the same y-axis
 /// - Consolidate sorted_metric_names across different runs, so that the frontend can know
 ///   what metric graphs to render as well the order of rendering
+/// - If all values of a metric are zero, compress the time-series values to reduce
+///   report data size
 fn post_process_time_series_data(processed_data: &mut ProcessedData) {
     let mut per_run_sorted_metric_names: Vec<&Vec<String>> = Vec::new();
     let mut per_metric_value_ranges: HashMap<String, Vec<(u64, u64)>> = HashMap::new();
@@ -194,7 +197,6 @@ fn post_process_time_series_data(processed_data: &mut ProcessedData) {
         per_metric_combined_value_range.insert(metric_name, combined_value_range);
     }
 
-    // Update the TimeSeriesData with the cross-run sorted metric names and value ranges
     for aperf_data in processed_data.runs.values_mut() {
         let time_series_data = match aperf_data {
             AperfData::TimeSeries(time_series_data) => time_series_data,
@@ -203,11 +205,16 @@ fn post_process_time_series_data(processed_data: &mut ProcessedData) {
                 return;
             }
         };
+        // Update the TimeSeriesData with the cross-run sorted metric names
         time_series_data.sorted_metric_names = sorted_metric_names.clone();
         for (metric_name, time_series_metric) in &mut time_series_data.metrics {
+            // Update every metric with cross-run combined value ranges
             if let Some(combined_value_range) = per_metric_combined_value_range.get(metric_name) {
                 time_series_metric.value_range = combined_value_range.to_owned();
             }
+            // Reduce the report data size for all-zero metrics by recreating a compressed
+            // series with only the first and last data points
+            compress_all_zero_time_series_metric(time_series_metric);
         }
     }
 }
